Switch to underscore scoping of nested object classes.
We've started to have collisions for some services (FooBar and Foo::Bar); and
the number seems to be increasing.

Add an option to disable it, we'll use that to not change the existing sources
that are already published.

While at it, ServiceGenerator always wants the SessionFetcher's logging to
support the --httpLogDir option.